dailybuild: suppress trivial warning
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@8252 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Samples/Physics/Reach/main.cpp
@@ -191,7 +191,7 @@ public:
 		// 初期視点
 		GetCurrentWin()->GetTrackball()->SetTarget  (Vec3f(0.0f, 0.2f, 0.0f));
 		GetCurrentWin()->GetTrackball()->SetDistance(2.0);
-		GetCurrentWin()->GetTrackball()->SetAngle   (Rad(135.0), Rad(30.0));
+		GetCurrentWin()->GetTrackball()->SetAngle   ((float)Rad(135.0), (float)Rad(30.0));
 	}
 
 	virtual void OnAction(int menu, int id){
